Debug

+++ server.py
@@ -110,11 +110,11 @@ class buyAdena(Resource):
                     cursor.execute("select count from items WHERE item_id=57 and owner_id=%s;", owner_id)
                     checkCurrentAdena = cursor.fetchall()
                     setAdenaFinal = (int(checkCurrentAdena[0]['count']) + count)
-                    adeptioToSet = float(float(count) / float(adeptio_BuyRate))
+                    adeptioToSet = int(int(count) / int(adeptio_BuyRate))
                     cursor.execute("update items set count=%s WHERE item_id=57 and owner_id=%s;", (setAdenaFinal, owner_id))
                     cursorLG.execute("select balance from adeptio_balances WHERE login=%s", account)
                     checkBalance = cursorLG.fetchall()
-                    setAdeptioFinal = float((float(checkBalance[0]['balance']) - float(adeptioToSet)))
+                    setAdeptioFinal = int((float(checkBalance[0]['balance']) - int(adeptioToSet)))
                     cursorLG.execute("replace into adeptio_balances (login, balance) values (%s, %s) ", (account, setAdeptioFinal))
                     return jsonify(data=success)
             else:
@@ -161,7 +161,7 @@ class sellAdena(Resource):
             if userCheck[0]['password'] == token:
                 if count and int(count) > 0:
                     setAdenaFinal = (int(adenaCountStatus[0]['count']) - count)
-                    adeptioTopay = float(count / adeptio_rate)
+                    adeptioTopay = int(count / adeptio_rate)
                     cursor.execute("update items set count=%s WHERE item_id=57 and owner_id=%s;", (setAdenaFinal, owner_id))
                     cursorLG.execute("select balance from adeptio_balances WHERE login=%s", account)
                     checkBalance = cursorLG.fetchall()
@@ -175,7 +175,7 @@ class sellAdena(Resource):
                     cursorLG.execute("select balance from adeptio_balances WHERE login=%s", account)
                     checkBalance = cursorLG.fetchall()
 
-                    setAdeptioFinal = float((int(checkBalance[0]['balance']) + float(adeptioTopay)))
+                    setAdeptioFinal = int((int(checkBalance[0]['balance']) + int(adeptioTopay)))
                     cursorLG.execute("replace into adeptio_balances (login, balance) values (%s, %s) ", (account, setAdeptioFinal))
                     return jsonify(data=success)
             else:
